+++ test/dummy/config/routes.rb
@@ -1,6 +1,8 @@
 Dummy::Application.routes.draw do
   devise_for :admins
   devise_for :users
+
+  devise_for :lockable_users
   devise_for :non_otp_users
 
   resources :posts